Updated Door
Updated the Door object to be a Door_Unit, which contains the sliding door and the door stop. Made this a prefab. The door slides up when being opened and locks in place. When closing gravity pulls the door back down until it collides with the door stop. If the door is locked, it cannot be opened or closed. This is controlled by the Door.cs script. I also made a DoorMenuButton script that is a modified version of Alex's DoorInteraction script to allow for the more complex interaction with the door. The new door can be found in Level_01 as well as the prefabs folder.
A new layer was made for the door so it doesn't collide with the walls and a new tag was made for objects the player cannot grab so that they can still try to interact with rigidbodies by pushing them but cannot grab and move them around.